requirements.txt
numpy==1.22.2
opencv-contrib-python-headless==4.5.5.62
opencv-python-headless==4.5.5.62
pdf2image==1.16.0
Pillow==9.0.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyYAML==6.0
imohash==1.0.4
pysmb==1.2.7